works using my development account only.
I also added a player object to experiment with what can be gotten from the player.
I also found that with the confluence skin, a video can be played in the background while doing other things.
Possibly can set the transparency of the pensive window pretty high, and watch the video while taking notes, same for audio or picture?